Found source of most bugs.  GNUstep really only works when the event loop is run on the main thread.  Not necessarily the thread where it is initialized, the main thread.
Using the trick to lock the OS thread in the init function for the module, which means that main runs on the main thread.  This works for applications, but does not work for testing.  The testing package owns the main thread, and uses goroutines for the tests.  This unfortunately means that the package cannot be tested.  Further investigation is required to find out how the testing package can be used.