Redesigned error handling for `testingTypeKeys`.
New design puts all errors into a channel.  This avoids any use of the
*testing.T from other goroutines, which was being done despite not
allowed by the API.

On windows, we check that there is a foreground window.  The reason is
unclear, but on newer CI machines there is no foreground window, which
completely breaks the SendInput functionality.  If the foreground window
is NULL, the test will be skipped