Added property tests for Update on several widgets
Added property tests (using quickcheck) to several basic widgets.  The
code is nearly identical to the same testing done for Buttons.

The timing for focus and blur tests was also adjusted.  The delays are a
little arbitrary, but testing code needs to wait for widgets to be
realized by the native system before focus can be taken.

Added a note that changing the field Password for TextInput widgets is
not supported on all platforms.