Fix bug with ampersand in text for WIN32 label
On windows, an ampersand in the text for a control is often used to
indicate a shortcut key.  For a static text control, this behaviour can
be disabled with the appropriate style (SS_NOPREFIX).

This commit does not fix similar bugs in other control, such as buttons
and checkboxes.  Those will require escaping the strings.