Bug fix in functions measuring scrollbars
The 2 functions replace the widget of the window with a scrollbar to
call functions to get the preferred dimensions.  When removed, the
old controls were being destroyed as the last reference was being
removed.  Fix was simply holding a reference to the widgets while the
scrollbar measurements are being done.