Cleanup of testing
Names of the widget testing functions have been updated to make clear
whether they do a test, or what type of help they provide.  Additional
use of the package goeytest to reuse the boilerplat in creating a window
and supplying an event loop so that a widget can be mounted.

Some layout bugs were uncovered by the additional testing, and have been
fixed.