Fixed test failures for Button on JS
The main window will now blur the active element when closing.  This
patches a platform difference, where this is done by the system on other
platforms.

Replaced use of innerText with textContent when accessing the label of
the Button.  This fixes some subtle bugs as innerText parse the text as
HTML.