Update Gitlab CI
Want verbose builds and tests, so that there is a log of which packages
were in fact built and tested.

Test matrices for JS/WASM and Cocoa are earliest version of Go
supported + latest.  Intermediate versions will not be tested.

Coverage profiles for 'go test' are currently not supported on JS/WASM.
Neither is the race detector.