Fix heisenbug in loop for windows
Although there weren't any panics or a faults, calls to loop.Do didn't
always synchronize correctly.  Updates to captures variables were not
always visible in the calling goroutine.  It is not clear why, but fix
involves an extra synchronization when sending the callback to the GUI
thread.

There was also an issue as the old code did not respect the rules for
uintptr.  Those could have been fixed, but the extra allocations are
unnecessary as all actions are serialized on the GUI thread anyway.
Replace the old code with globals, which avoids the need for the uintptr
sublety.