Fix intermittent error in calls to CreateBItmap
Calls to CreateBitmap on windows, as well as to GetDIBits, suffered from
intermitten failures.  The error returned is invalid argument.  Testing
found that the failures occur when the memory block passed for pixel
data includes 0xc000400000.  This pre-condition does not appear to be
documented.  The reason for the restriction is unclear.

The current fix is to create a new block, copying data if necessary.

There may be a performance improvement in using a sync.Pool for the
internal memory blocks used when converting image.Image to HBITMAPs.