Clean up of loop.go
This work is to support tinygo, where the calls to runtime.LockOSThread
are not supported.

This commit also fixes a bug intnroduced by the latest docker image for
NixOS.  Apparently, these images no longer have a /etc/*-release file.
There was also changed behaviour with respect to CA certs, and directory
handling in nix-shell.